3/12

+++ app/src/main/java/ru/bacha/gym/Approach.java
@@ -7,13 +7,13 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Approach {
     @PrimaryKey
-    public int id;
+    public Long id;
     public double weight;
     public int replay;
     public String note;
 
 
-    public Approach(int id, double weight, int replay, String note) {
+    public Approach(Long id, double weight, int replay, String note) {
         this.id = id;
         this.weight = weight;
         this.replay = replay;
